cformer: add basic layer framework
Layer is the basic unit of neural network. It can be viewd as collection of neurons
and the connections(a.k.a parameters) between them. On top of layers, we can build
complex neural network, which will be implemented later.

This patch adds the basic layer construct and implement a simple linear layer.

- add xavier and kaiming initiliazer

Signed-off-by: Liu Yuan <namei.unix@gmail.com>