tensor: fix log(0)
ReLU will make some elements in the tensor 0 and feed to cross entropy
loss function, which will make the log(0) in the loss function. The log(0)
is -inf, so we replace 0 with a small value to make log(0) a small negative
number.

Pytorch and tensorflow do the same thing in their cross entropy loss funcs.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>