tensor: fix bwd_log NaN problem
The problem is that softmax of a very confident network will return Extreamly
Small Number(ESN) or even 0 (if underflow) for the correct class. This will
cause log(0) and make gradient as Extreamly Huge Number(EHN) or NaN.

For e.g,
{1.0000, ESN, ...} y_pred.data
{0, 1, ...} y_true.data
{0, EHN, ...} log(y_pred)).grad

or
{1.000, 0, ... } y_pred.data
{0, 1, ...} y_true.data
{0, NaN, ...} log(y_pred)).grad

With lr=5e-4, we can reproduce this problem easily.

This patch fix this problem by clipping the y_pred.data to a very small number.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>